mcp-postgres-server/pyproject.toml
[project]
name = "mcp-postgres-server"
version = "0.1.0"
description = "MCP server for PostgreSQL database operations - Spartans Hockey"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "asyncpg>=0.29.0",
  "fastmcp>=2.9.0",
  "python-dotenv>=1.0.0"
]

[tool.uv]
dev-dependencies = []
